backend/requirements.txt
flask
tflite-runtime==2.14.0
numpy==1.26.4
pillow
gunicorn